fix(security): bind dashboards to localhost, sanitize session/project path components
- SEC-1: research + eval dashboards bound 0.0.0.0 but logged 'localhost' (misleading).
  Bind explicit 127.0.0.1 by default, log the actual address; opt-in host override.
- SEC-2: sanitize session_id before it becomes a path component (getSessionId chokepoint
  + getDiaryDraftPath/getDiaryPath); fail-safe to a hashed id, never throws (hooks tier).
- SEC-3: getProjectName() sanitizes the basename, mirroring the worktree layer. Consolidated
  the duplicated sanitizeProjectName into canonical utils.js; worktree-paths.js now imports
  it (no logic duplication, per architecture.md canonical-source rule).

Defense-in-depth: session_id/project name are not model-controlled today, but this closes
the path-escape surface and removes the misleading bind log.
